Core/Player: Extend action button value to uint64 to be able to hold battle pet guids
Port From (https://github.com/TrinityCore/TrinityCore/commit/f67cd38312014b13624dcb5fe1d117dac4892b7d)

+++ Source/Game/Handlers/MiscHandler.cs
@@ -113,7 +113,7 @@ namespace Game
         [WorldPacketHandler(ClientOpcodes.SetActionButton)]
         void HandleSetActionButton(SetActionButton packet)
         {
-            uint action = packet.GetButtonAction();
+            ulong action = packet.GetButtonAction();
             uint type = packet.GetButtonType();
 
             if (packet.Action == 0)
